pharmacyhelp/urls.py
from django.conf.urls import url, include
from django.contrib import admin
from drf_yasg import openapi
from drf_yasg.generators import OpenAPISchemaGenerator
from drf_yasg.views import get_schema_view
from rest_framework import permissions

schema_view = get_schema_view(
   openapi.Info(
      title="Snippets API",
      default_version='v1',
      description="Test description",
      terms_of_service="https://www.google.com/policies/terms/",
      contact=openapi.Contact(email="contact@snippets.local"),
      license=openapi.License(name="BSD License"),
   ),
   public=True,
   permission_classes=(permissions.AllowAny,),
)

urlpatterns = [
    url('admin/', admin.site.urls),
    url('docs/', schema_view.with_ui('swagger', cache_timeout=0)),
    url('', include('recipes.urls')),
]
