requirements.txt
django==2.2.4
djangorestframework==3.10.2
django-cors-headers==3.1.0
drf-yasg==1.16.1
packaging==19.1
